attempt recognition of Inform by GitHub
This addition to the dot-gitattributes file causes the Linguist
tool, which GH uses to analyze determine repo's language, to only pay
attention to files with inform and ni extensions. If Linguist knows
about Inform at all, this should make it recognize it; if not then oh
well. (Not important enough to download Linguist myself and verify, I
have things to do today.)