COMMIT THE FINAL SOURCE TO VERSION CONTROL
There are huge changes here, including ones made to core areas of
the game. This shows that the last commit for this file  was very
early on in the game design process.

Some of the things included in this mega-commit are:
- all the train code, replacing the old timer system
- the credits command
- the directions command
- the list of exits on the left hand status line
- all the code for gameplay inside the fake tree
- the ending poem
- at least half of the non-speech Ana code
table
- all of the conversation code and everyone's conversation tables
- lots of details for bird interaction

Amazingly, all of that made it into the game AFTER the commissary.