allow frying pan to hold objects
In the course of adding this code, I also reordered the "instead"
rules under "check cooking something" so that they go from least to
most specific. Previously, trying to cook something that wasn't food
would trigger the "if the munch is not a food" rule, even if the
player wasn't in the right location and didn't have the pan. That's a
bug -- and now it's fixed.