add "distant island" backdrop wherever it should be visible
Adding this backdrop to these locations makes "x/look island"
possible.

The player expects that to be possible since they have already seen
the island from the apartment window.

In order to make this work, I had to go through and change all
references to the Island of Technique to use its full name, so that
references to "the island", meaning IoT, did not collide with
references to "the island" meaning this new backdrop.

Otherwise, Inform thought that "distant island" referred to the IoT,
which (being a room) is a container, and therefore couldn't also be a backdrop.