add release 3 story file
Unless I am sorely mistaken, the only gameplay changes here is in the
money-related code "now the cashola" and "if a random chance of 1 in
4".

These two clauses fix the implementation of money so that it reappears
in the commissary. The removal of comments on line ~913 is because of this change.